Move assign_custom_measurable_outputs to aeppl.abstract
